Inline PCA and LDA JSON serialization
